LayoutTests/platform/glib/imported/w3c/web-platform-tests/server-timing/cross_origin-expected.txt


PASS Expected entry count for http://web-platform.test:8800/server-timing/resources/blue.png: 1
PASS Expected entry count for http://www1.web-platform.test:8800/server-timing/resources/blue.png: 0
FAIL Expected entry count for http://www1.web-platform.test:8800/server-timing/resources/blue_tao.png: 1 assert_equals: Expected entry count for http://www1.web-platform.test:8800/server-timing/resources/blue_tao.png: 1 expected 1 but got 0
